panel: split menu widget files up
