--- src/panel/widgets/menu.hpp
@@ -1,231 +1,0 @@
-#pragma once
-
-#include <giomm/desktopappinfo.h>
-#include <gtkmm.h>
-#include <sigc++/connection.h>
-#include <set>
-
-#include "widget.hpp"
-#include "wf-popover.hpp"
-
-class WayfireMenu;
-using AppInfo = Glib::RefPtr<Gio::DesktopAppInfo>;
-
-class WfMenuLayout : public Gtk::LayoutManager
-{
-  protected:
-    Gtk::Allocation search_alloc, logout_alloc, category_alloc, flow_alloc, separator_alloc;
-
-    void allocate_vfunc(const Gtk::Widget& widget, int width, int height, int baseline) override;
-    void measure_vfunc(const Gtk::Widget& widget, Gtk::Orientation orientation, int for_size, int& minimum,
-        int& natural, int& minimum_baseline, int& natural_baseline) const override;
-    WayfireMenu *menu;
-    int limit_width = 0, limit_height = 0;
-
-    WfOption<bool> show_categories{"panel/menu_show_categories"};
-    WfOption<int> category_width{"panel/menu_min_category_width"};
-    WfOption<int> content_width{"panel/menu_min_content_width"};
-    WfOption<int> content_height{"panel/menu_min_content_height"};
-    WfOption<std::string> panel_position{"panel/position"};
-
-  public:
-    WfMenuLayout(WayfireMenu *menu);
-
-    void set_limit(int x, int y);
-};
-
-
-class WfMenuCategory
-{
-  public:
-    WfMenuCategory(std::string name, std::string icon_name);
-    std::string get_name();
-    std::string get_icon_name();
-    std::vector<AppInfo> items;
-
-  private:
-    std::string name;
-    std::string icon_name;
-};
-
-class WfMenuCategoryButton : public Gtk::Button
-{
-  public:
-    WfMenuCategoryButton(WayfireMenu *menu, std::string category, std::string label, std::string icon_name);
-    ~WfMenuCategoryButton();
-
-  private:
-    WayfireMenu *menu;
-    Gtk::Box m_box;
-    Gtk::Label m_label;
-    Gtk::Image m_image;
-    sigc::connection sig_click;
-
-    std::string category;
-    std::string label;
-    std::string icon_name;
-    void on_click();
-};
-
-class WfMenuItem : public Gtk::FlowBoxChild
-{
-  public:
-    WfMenuItem(WayfireMenu *menu, AppInfo app);
-    ~WfMenuItem();
-
-    uint32_t matches(Glib::ustring text);
-    uint32_t fuzzy_match(Glib::ustring text);
-    bool operator <(const WfMenuItem& other);
-    void set_search_value(uint32_t value);
-    uint32_t get_search_value();
-    void on_click();
-
-  private:
-    WayfireMenu *menu;
-    Gtk::Box box, list_item;
-    Gtk::Image image;
-    Gtk::Label label;
-    Glib::RefPtr<Gio::Menu> m_menu;
-    Glib::RefPtr<Gio::SimpleActionGroup> actions;
-    Gtk::Button button;
-    Gtk::MenuButton extra_actions_button;
-    std::vector<sigc::connection> signals;
-
-    bool has_actions = false;
-    uint32_t search_value = 0;
-
-    AppInfo app_info;
-};
-
-class WayfireLogoutUIButton
-{
-  public:
-    Gtk::Box layout;
-    Gtk::Image image;
-    Gtk::Label label;
-    Gtk::Button button;
-};
-
-class WayfireLogoutUI
-{
-  public:
-    WayfireLogoutUI();
-    ~WayfireLogoutUI();
-    WfOption<std::string> logout_command{"panel/logout_command"};
-    WfOption<std::string> reboot_command{"panel/reboot_command"};
-    WfOption<std::string> shutdown_command{"panel/shutdown_command"};
-    WfOption<std::string> suspend_command{"panel/suspend_command"};
-    WfOption<std::string> hibernate_command{"panel/hibernate_command"};
-    WfOption<std::string> switchuser_command{"panel/switchuser_command"};
-    Gtk::Window ui;
-    WayfireLogoutUIButton logout;
-    WayfireLogoutUIButton reboot;
-    WayfireLogoutUIButton shutdown;
-    WayfireLogoutUIButton suspend;
-    WayfireLogoutUIButton hibernate;
-    WayfireLogoutUIButton switchuser;
-    WayfireLogoutUIButton cancel;
-    Gtk::CenterBox box;
-    Gtk::Grid main_layout;
-    std::vector<sigc::connection> signals;
-    void create_logout_ui_button(WayfireLogoutUIButton *button,
-        const char *icon, const char *label);
-    void on_logout_click();
-    void on_reboot_click();
-    void on_shutdown_click();
-    void on_suspend_click();
-    void on_hibernate_click();
-    void on_switchuser_click();
-    void on_cancel_click();
-};
-
-
-class WayfireMenu : public WayfireWidget
-{
-    WayfireOutput *output;
-
-  public:
-    Gtk::Box flowbox_container;
-    Gtk::Box box, box_bottom;
-    Gtk::Box popover_layout_box;
-    Gtk::Box category_box;
-    Gtk::Separator separator;
-    Gtk::Image main_image;
-    Gtk::SearchEntry search_entry;
-    Gtk::FlowBox flowbox;
-    Gtk::Button logout_button;
-    Gtk::Image logout_image;
-    Gtk::ScrolledWindow app_scrolled_window, category_scrolled_window;
-
-  private:
-    std::shared_ptr<WfMenuLayout> layout;
-    std::unique_ptr<WayfireMenuWidget> button;
-    std::unique_ptr<WayfireLogoutUI> logout_ui;
-
-    GAppInfoMonitor *app_info_monitor = nullptr;
-    guint app_info_monitor_changed_handler_id;
-
-    void load_menu_item(AppInfo app_info);
-    void load_menu_items_from_dir(std::string directory);
-    void load_menu_items_all();
-
-    void add_category_app(std::string category, Glib::RefPtr<Gio::DesktopAppInfo>);
-
-    bool update_icon();
-
-    bool m_sort_names     = true;
-    bool fuzzy_filter     = false;
-    int32_t count_matches = 0;
-
-    bool on_sort(Gtk::FlowBoxChild*, Gtk::FlowBoxChild*);
-    bool on_filter(Gtk::FlowBoxChild *child);
-    void on_search_changed();
-    void on_popover_shown();
-
-    /* loaded_apps is a list of the already-opened applications + their execs,
-     * so that we don't show duplicate entries */
-    std::set<std::pair<std::string, std::string>> loaded_apps;
-    std::unordered_map<std::string, std::unique_ptr<WfMenuCategory>> category_list;
-    std::string category = "All";
-    std::vector<std::string> category_order = {
-        "All", "Network", "Education", "Office", "Development", "Graphics", "AudioVideo", "Game", "Science",
-        "Settings", "System", "Utility", "Hidden"
-    };
-
-    std::vector<sigc::connection> signals;
-
-    WfOption<std::string> menu_logout_command{"panel/menu_logout_command"};
-    WfOption<bool> fuzzy_search_enabled{"panel/menu_fuzzy_search"};
-    WfOption<std::string> panel_position{"panel/position"};
-    WfOption<std::string> menu_icon{"panel/menu_icon"};
-    WfOption<int> flowbox_spacing{"panel/menu_item_spacing"};
-    WfOption<int> menu_min_category_width{"panel/menu_min_category_width"};
-    WfOption<int> menu_min_content_height{"panel/menu_min_content_height"};
-    WfOption<bool> menu_show_categories{"panel/menu_show_categories"};
-    void setup_popover_layout();
-    void update_popover_layout();
-    void update_size();
-    void create_logout_ui();
-    void on_logout_click();
-    void key_press_search();
-    void select_first_flowbox_item();
-
-  public:
-    void arrow_key(Gtk::DirectionType dir);
-    void init(Gtk::Box *container) override;
-    void populate_menu_items(std::string category);
-    void populate_menu_categories();
-    void toggle_menu();
-    void hide_menu();
-    void refresh();
-    void set_category(std::string category);
-    WfOption<bool> menu_list{"panel/menu_list"};
-    WfOption<int> menu_min_content_width{"panel/menu_min_content_width"};
-
-    WayfireMenu(WayfireOutput *output)
-    {
-        this->output = output;
-    }
-
-    ~WayfireMenu() override;
-};
